Removed nasty warning with javac

+++ core/src/main/java/com/orientechnologies/orient/core/metadata/schema/OClass.java
@@ -254,7 +254,7 @@ public class OClass extends ODocumentWrapperNoClass implements Comparable<OClass
 		id = (Integer) document.field("id");
 		defaultClusterId = (Integer) document.field("defaultClusterId");
 
-		overSize = document.field("overSize");
+		overSize = (Float) document.field("overSize");
 
 		final Object cc = document.field("clusterIds");
 		if (cc instanceof Collection<?>) {
